Move income model and data from conftest.py to test_white_box_sklearn_cases.py
