+++ runtimes/alibi-explain/tests/conftest.py
@@ -15,8 +15,6 @@ from httpx import AsyncClient
 from fastapi import FastAPI
 from prometheus_client.registry import REGISTRY, CollectorRegistry
 from starlette_exporter import PrometheusMiddleware
-from sklearn.ensemble import GradientBoostingClassifier
-from alibi.datasets import fetch_adult
 from alibi.api.interfaces import Explanation, Explainer
 from alibi.explainers import AnchorImage
 
@@ -326,36 +324,3 @@ def _train_anchor_image_explainer() -> None:
 
     _ANCHOR_IMAGE_DIR.mkdir(parents=True)
     anchor_image.save(_ANCHOR_IMAGE_DIR)
-
-
-@pytest.fixture(scope="function")  # must scope as function due to use of tmp_path
-def sk_income_model_uri(income_data, tmp_path) -> str:
-    X_train, Y_train = income_data['X'], income_data['Y']
-
-    model = GradientBoostingClassifier(n_estimators=50)
-    model.fit(X_train, Y_train)
-
-    model_uri = os.path.join(tmp_path, "sklearn-model.joblib")
-    joblib.dump(model, model_uri)
-
-    return model_uri
-
-
-@pytest.fixture(scope="module")
-def income_data() -> dict:
-    adult = fetch_adult()
-    X = adult.data
-    Y = adult.target
-
-    feature_names = adult.feature_names
-    category_map = adult.category_map
-
-    # Package into dictionary
-    data_dict = {
-        'X': X,
-        'Y': Y,
-        'feature_names': feature_names,
-        'category_map': category_map,
-        'target_names': adult.target_names,
-    }
-    return data_dict
