+++ runtimes/alibi-explain/tests/test_white_box_sklearn_cases.py
@@ -1,11 +1,48 @@
 from typing import Tuple
-
+import os
+import joblib
 import pytest
+
 from alibi.api.interfaces import Explainer
+from sklearn.ensemble import GradientBoostingClassifier
+from alibi.datasets import fetch_adult
 
 from mlserver.types import InferenceRequest
 from mlserver.settings import ModelSettings
 from .helpers.utils import build_test_case
+
+
+@pytest.fixture(scope="function")  # must scope as function due to use of tmp_path
+def sk_income_model_uri(income_data, tmp_path) -> str:
+    X_train, Y_train = income_data['X'], income_data['Y']
+
+    model = GradientBoostingClassifier(n_estimators=50)
+    model.fit(X_train, Y_train)
+
+    model_uri = os.path.join(tmp_path, "sklearn-model.joblib")
+    joblib.dump(model, model_uri)
+
+    return model_uri
+
+
+@pytest.fixture(scope="module")
+def income_data() -> dict:
+    adult = fetch_adult()
+    X = adult.data
+    Y = adult.target
+
+    feature_names = adult.feature_names
+    category_map = adult.category_map
+
+    # Package into dictionary
+    data_dict = {
+        'X': X,
+        'Y': Y,
+        'feature_names': feature_names,
+        'category_map': category_map,
+        'target_names': adult.target_names,
+    }
+    return data_dict
 
 
 def case_tree_shap(sk_income_model_uri, income_data, tmp_path) \
